+++ tokora/src/parser/many/sep/delim/mod.rs
@@ -252,6 +252,7 @@ impl<'inp, L, P, Sep, O, Ctx, Delim, Lang: ?Sized, Cmpl>
             .emitter()
             .emit_unclosed(Unclosed::<Delim, L::Span, Lang>::of(
               open_span,
+              Delim::KIND,
               Delim::name(),
             ))?;
         }